refactor: move confirm code to store
